undo fmt
