+++ taichi/analysis/detect_fors_with_break.cpp
@@ -1,4 +1,5 @@
 #include "taichi/ir/ir.h"
+#include "taichi/ir/frontend_ir.h"
 
 #include <unordered_set>
