+++ taichi/ir/basic_stmt_visitor.cpp
@@ -1,4 +1,5 @@
-#include "visitors.h"
+#include "taichi/ir/visitors.h"
+#include "taichi/ir/frontend_ir.h"
 
 TLANG_NAMESPACE_BEGIN
