+++ taichi/ir/ir.cpp
@@ -1005,4 +1005,34 @@ bool ContinueStmt::as_return() const {
   return false;
 }
 
+If::If(const Expr &cond) {
+  auto stmt_tmp = std::make_unique<FrontendIfStmt>(cond);
+  stmt = stmt_tmp.get();
+  current_ast_builder().insert(std::move(stmt_tmp));
+}
+
+If::If(const Expr &cond, const std::function<void()> &func) : If(cond) {
+  Then(func);
+}
+
+If &If::Then(const std::function<void()> &func) {
+  auto _ = current_ast_builder().create_scope(stmt->true_statements);
+  func();
+  return *this;
+}
+
+If &If::Else(const std::function<void()> &func) {
+  auto _ = current_ast_builder().create_scope(stmt->false_statements);
+  func();
+  return *this;
+}
+
+While::While(const Expr &cond, const std::function<void()> &func) {
+  auto while_stmt = std::make_unique<FrontendWhileStmt>(cond);
+  FrontendWhileStmt *ptr = while_stmt.get();
+  current_ast_builder().insert(std::move(while_stmt));
+  auto _ = current_ast_builder().create_scope(ptr->body);
+  func();
+}
+
 TLANG_NAMESPACE_END
