+++ taichi/ir/ir.h
@@ -953,24 +953,6 @@ class ArgLoadExpression : public Expression {
 };
 
 // For return values
-class FrontendArgStoreStmt : public Stmt {
- public:
-  int arg_id;
-  Expr expr;
-
-  FrontendArgStoreStmt(int arg_id, const Expr &expr)
-      : arg_id(arg_id), expr(expr) {
-  }
-
-  // Arguments are considered global (nonlocal)
-  virtual bool has_global_side_effect() const override {
-    return true;
-  }
-
-  DEFINE_ACCEPT
-};
-
-// For return values
 class ArgStoreStmt : public Stmt {
  public:
   int arg_id;
@@ -1325,21 +1307,6 @@ class Block : public IRNode {
   DEFINE_ACCEPT
 };
 
-class FrontendSNodeOpStmt : public Stmt {
- public:
-  SNodeOpType op_type;
-  SNode *snode;
-  ExprGroup indices;
-  Expr val;
-
-  FrontendSNodeOpStmt(SNodeOpType op_type,
-                      SNode *snode,
-                      const ExprGroup &indices,
-                      const Expr &val = Expr(nullptr));
-
-  DEFINE_ACCEPT
-};
-
 class SNodeOpStmt : public Stmt {
  public:
   SNodeOpType op_type;
@@ -1363,18 +1330,6 @@ class SNodeOpStmt : public Stmt {
   }
 
   TI_STMT_DEF_FIELDS(ret_type, op_type, snode, ptr, val, indices);
-  DEFINE_ACCEPT
-};
-
-class FrontendAssertStmt : public Stmt {
- public:
-  std::string text;
-  Expr val;
-
-  FrontendAssertStmt(const std::string &text, const Expr &val)
-      : text(text), val(val) {
-  }
-
   DEFINE_ACCEPT
 };
 
@@ -1413,15 +1368,6 @@ class RangeAssumptionStmt : public Stmt {
   }
 
   TI_STMT_DEF_FIELDS(ret_type, input, base, low, high);
-  DEFINE_ACCEPT
-};
-
-class FrontendAssignStmt : public Stmt {
- public:
-  Expr lhs, rhs;
-
-  FrontendAssignStmt(const Expr &lhs, const Expr &rhs);
-
   DEFINE_ACCEPT
 };
 
@@ -1525,44 +1471,6 @@ class IfStmt : public Stmt {
   DEFINE_ACCEPT
 };
 
-class FrontendIfStmt : public Stmt {
- public:
-  Expr condition;
-  std::unique_ptr<Block> true_statements, false_statements;
-
-  FrontendIfStmt(const Expr &condition) : condition(load_if_ptr(condition)) {
-  }
-
-  bool is_container_statement() const override {
-    return true;
-  }
-
-  DEFINE_ACCEPT
-};
-
-class FrontendPrintStmt : public Stmt {
- public:
-  Expr expr;
-  std::string str;
-
-  FrontendPrintStmt(const Expr &expr, const std::string &str)
-      : expr(load_if_ptr(expr)), str(str) {
-  }
-
-  DEFINE_ACCEPT
-};
-
-class FrontendEvalStmt : public Stmt {
- public:
-  Expr expr;
-  Expr eval_expr;
-
-  FrontendEvalStmt(const Expr &expr) : expr(load_if_ptr(expr)) {
-  }
-
-  DEFINE_ACCEPT
-};
-
 class PrintStmt : public Stmt {
  public:
   Stmt *stmt;
@@ -1580,27 +1488,13 @@ class If {
  public:
   FrontendIfStmt *stmt;
 
-  If(const Expr &cond) {
-    auto stmt_tmp = std::make_unique<FrontendIfStmt>(cond);
-    stmt = stmt_tmp.get();
-    current_ast_builder().insert(std::move(stmt_tmp));
-  }
+  explicit If(const Expr &cond);
 
-  If(const Expr &cond, const std::function<void()> &func) : If(cond) {
-    Then(func);
-  }
+  If(const Expr &cond, const std::function<void()> &func);
 
-  If &Then(const std::function<void()> &func) {
-    auto _ = current_ast_builder().create_scope(stmt->true_statements);
-    func();
-    return *this;
-  }
+  If &Then(const std::function<void()> &func);
 
-  If &Else(const std::function<void()> &func) {
-    auto _ = current_ast_builder().create_scope(stmt->false_statements);
-    func();
-    return *this;
-  }
+  If &Else(const std::function<void()> &func);
 };
 
 class ConstStmt : public Stmt {
@@ -1628,37 +1522,6 @@ class ConstStmt : public Stmt {
   std::unique_ptr<ConstStmt> copy();
 
   TI_STMT_DEF_FIELDS(ret_type, val);
-  DEFINE_ACCEPT
-};
-
-class FrontendForStmt : public Stmt {
- public:
-  Expr begin, end;
-  Expr global_var;
-  std::unique_ptr<Block> body;
-  std::vector<Identifier> loop_var_id;
-  int vectorize;
-  int parallelize;
-  bool strictly_serialized;
-  ScratchPadOptions scratch_opt;
-  int block_dim;
-
-  bool is_ranged() const {
-    if (global_var.expr == nullptr) {
-      return true;
-    } else {
-      return false;
-    }
-  }
-
-  FrontendForStmt(const ExprGroup &loop_var, const Expr &global_var);
-
-  FrontendForStmt(const Expr &loop_var, const Expr &begin, const Expr &end);
-
-  bool is_container_statement() const override {
-    return true;
-  }
-
   DEFINE_ACCEPT
 };
 
@@ -1767,21 +1630,6 @@ class FuncBodyStmt : public Stmt {
   DEFINE_ACCEPT
 };
 
-class FrontendFuncDefStmt : public Stmt {
- public:
-  std::string funcid;
-  std::unique_ptr<Block> body;
-
-  FrontendFuncDefStmt(const std::string &funcid) : funcid(funcid) {
-  }
-
-  bool is_container_statement() const override {
-    return true;
-  }
-
-  DEFINE_ACCEPT
-};
-
 class FuncCallStmt : public Stmt {
  public:
   std::string funcid;
@@ -1813,44 +1661,6 @@ class WhileStmt : public Stmt {
   }
 
   TI_STMT_DEF_FIELDS(mask);
-  DEFINE_ACCEPT
-};
-
-class FrontendBreakStmt : public Stmt {
- public:
-  FrontendBreakStmt() {
-  }
-
-  bool is_container_statement() const override {
-    return false;
-  }
-
-  DEFINE_ACCEPT
-};
-
-class FrontendContinueStmt : public Stmt {
- public:
-  FrontendContinueStmt() = default;
-
-  bool is_container_statement() const override {
-    return false;
-  }
-
-  DEFINE_ACCEPT
-};
-
-class FrontendWhileStmt : public Stmt {
- public:
-  Expr cond;
-  std::unique_ptr<Block> body;
-
-  FrontendWhileStmt(const Expr &cond) : cond(load_if_ptr(cond)) {
-  }
-
-  bool is_container_statement() const override {
-    return true;
-  }
-
   DEFINE_ACCEPT
 };
 
@@ -2043,13 +1853,7 @@ class For {
 
 class While {
  public:
-  While(const Expr &cond, const std::function<void()> &func) {
-    auto while_stmt = std::make_unique<FrontendWhileStmt>(cond);
-    FrontendWhileStmt *ptr = while_stmt.get();
-    current_ast_builder().insert(std::move(while_stmt));
-    auto _ = current_ast_builder().create_scope(ptr->body);
-    func();
-  }
+  While(const Expr &cond, const std::function<void()> &func);
 };
 
 Expr Var(const Expr &x);
