+++ taichi/transforms/reverse_segments.cpp
@@ -1,4 +1,6 @@
 #include "taichi/ir/ir.h"
+#include "taichi/ir/frontend_ir.h"
+
 #include <set>
 
 TLANG_NAMESPACE_BEGIN
